Refactor: 이미지 객체를 임베디드 타입으로 사용하도록 변경
